+++ src/main/java/chess/database/DBConnector.java
@@ -33,4 +33,14 @@ public class DBConnector {
 
         return conn;
     }
+
+    // 드라이버 연결해제
+    public void closeConnection(Connection con) {
+        try {
+            if (con != null)
+                con.close();
+        } catch (SQLException e) {
+            System.err.println("con 오류:" + e.getMessage());
+        }
+    }
 }
